// web/src/api/models/graph-edge.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GraphEdgeProperties } from './graph-edge-properties';

/**
 * Knowledge graph edge representing a relationship
 * @export
 * @interface GraphEdge
 */
export interface GraphEdge {
    /**
     * Unique identifier for the edge
     * @type {string}
     * @memberof GraphEdge
     */
    'id': string;
    /**
     * Type of the relationship
     * @type {string}
     * @memberof GraphEdge
     */
    'type'?: string;
    /**
     * Source node ID
     * @type {string}
     * @memberof GraphEdge
     */
    'source': string;
    /**
     * Target node ID
     * @type {string}
     * @memberof GraphEdge
     */
    'target': string;
    /**
     * 
     * @type {GraphEdgeProperties}
     * @memberof GraphEdge
     */
    'properties': GraphEdgeProperties;
}