// web/src/api/models/graph-node.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GraphNodeProperties } from './graph-node-properties';

/**
 * Knowledge graph node representing an entity
 * @export
 * @interface GraphNode
 */
export interface GraphNode {
    /**
     * Unique identifier for the node (entity name)
     * @type {string}
     * @memberof GraphNode
     */
    'id': string;
    /**
     * Labels associated with the node
     * @type {Array<string>}
     * @memberof GraphNode
     */
    'labels': Array<string>;
    /**
     * 
     * @type {GraphNodeProperties}
     * @memberof GraphNode
     */
    'properties': GraphNodeProperties;
}