// web/src/api/models/knowledge-graph.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { GraphEdge } from './graph-edge';
// May contain unused imports in some cases
// @ts-ignore
import type { GraphNode } from './graph-node';

/**
 * Knowledge graph containing nodes and edges
 * @export
 * @interface KnowledgeGraph
 */
export interface KnowledgeGraph {
    /**
     * List of nodes in the knowledge graph
     * @type {Array<GraphNode>}
     * @memberof KnowledgeGraph
     */
    'nodes': Array<GraphNode>;
    /**
     * List of edges in the knowledge graph
     * @type {Array<GraphEdge>}
     * @memberof KnowledgeGraph
     */
    'edges': Array<GraphEdge>;
    /**
     * Whether the graph was truncated due to size limits
     * @type {boolean}
     * @memberof KnowledgeGraph
     */
    'is_truncated': boolean;
}